# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.98:8280/api",
            "description": "Сервер"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/worksheet/modules/credits": {
            "get": {
                "tags": [
                    "МОДУЛЬ РЛ: Кредит"
                ],
                "summary": "Список для рабочего листа (worksheet_id = int) ",
                "description": "Список для рабочего листа ",
                "operationId": "creditList",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "МОДУЛЬ РЛ: Кредит"
                ],
                "summary": "Создать",
                "description": "Создать",
                "operationId": "creditPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreditCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/worksheet/modules/credits/{id}": {
            "get": {
                "tags": [
                    "МОДУЛЬ РЛ: Кредит"
                ],
                "summary": "Открыть",
                "description": "Открыть",
                "operationId": "creditShow",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "МОДУЛЬ РЛ: Кредит"
                ],
                "summary": "Удалить",
                "description": "Удалить",
                "operationId": "creditDelete",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "МОДУЛЬ РЛ: Кредит"
                ],
                "summary": "Изменить",
                "description": "Изменить",
                "operationId": "creditPatch",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreditCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/worksheet/modules/services": {
            "get": {
                "tags": [
                    "МОДУЛЬ РЛ: Финансовые сервисы"
                ],
                "summary": "Список для рабочего листа (worksheet_id = int) ",
                "description": "Список для рабочего листа ",
                "operationId": "worksheetServiceList",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "МОДУЛЬ РЛ: Финансовые сервисы"
                ],
                "summary": "Создать ",
                "description": "Создать ",
                "operationId": "worksheetServiceCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorksheetServiceCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/worksheet/modules/services/{id}": {
            "get": {
                "tags": [
                    "МОДУЛЬ РЛ: Финансовые сервисы"
                ],
                "summary": "Открыть ",
                "description": "Открыть ",
                "operationId": "worksheetServiceShow",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "МОДУЛЬ РЛ: Финансовые сервисы"
                ],
                "summary": "Удалить тест",
                "description": "Удалить тест",
                "operationId": "worksheetServiceDestroy",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "МОДУЛЬ РЛ: Финансовые сервисы"
                ],
                "summary": "Изменить ",
                "description": "Изменить ",
                "operationId": "worksheetServiceUpdate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorksheetServiceCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreditCreateRequest": {
                "properties": {
                    "worksheet": {
                        "description": "ID рабочего листа",
                        "type": "integer",
                        "format": "integer"
                    },
                    "debtor": {
                        "description": "ID заемщика",
                        "type": "integer",
                        "format": "integer"
                    },
                    "tactic": {
                        "description": "ID тактики расчета",
                        "type": "integer",
                        "format": "integer"
                    },
                    "creditor": {
                        "description": "ID кредитора",
                        "type": "integer",
                        "format": "integer"
                    },
                    "status": {
                        "description": "ID статуса заявки",
                        "type": "integer",
                        "format": "integer"
                    },
                    "author": {
                        "description": "ID автора заявки",
                        "type": "integer",
                        "format": "integer"
                    },
                    "close": {
                        "description": "расторжение",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "period": {
                        "description": "Срок",
                        "type": "integer",
                        "format": "integer"
                    },
                    "cost": {
                        "description": "Стоимость",
                        "type": "integer",
                        "format": "integer"
                    },
                    "first_pay": {
                        "description": "ПВ",
                        "type": "integer",
                        "format": "integer"
                    },
                    "month_pay": {
                        "description": "Месячный платеж",
                        "type": "integer",
                        "format": "integer"
                    },
                    "simple": {
                        "description": "Упрощнный растет",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "approximates": {
                        "description": "Наполнение (примерное) те сервисы (именно сервисы, а не модули РЛ), которые возможно добавятся (api/finservices)",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": [
                                1,
                                2
                            ]
                        }
                    },
                    "register_at": {
                        "description": "Дата оформления",
                        "type": "string",
                        "format": "string"
                    },
                    "decorator": {
                        "description": "Оформитель",
                        "type": "integer",
                        "format": "integer"
                    },
                    "award": {
                        "description": "Награда",
                        "type": "integer",
                        "format": "integer"
                    },
                    "award_complete": {
                        "description": "Начисление награды",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "deduction": {
                        "description": "удержание",
                        "type": "integer",
                        "format": "integer"
                    },
                    "services": {
                        "description": "Модули финансовых услуг (те финуслуги что есть в рл)",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": [
                                1,
                                2
                            ]
                        }
                    },
                    "car": {
                        "description": "Машина",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": {
                                "id": "2",
                                "type": "new"
                            }
                        }
                    }
                },
                "type": "object"
            },
...
            "WorksheetServiceCreateRequest": {
                "properties": {
                    "car": {
                        "description": "Машина",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": {
                                "id": "2",
                                "type": "new"
                            }
                        }
                    },
                    "service": {
                        "description": "ID финансовой услуги",
                        "type": "integer",
                        "format": "integer"
                    },
                    "provider": {
                        "description": "ID провайдера",
                        "type": "integer",
                        "format": "integer"
                    },
                    "cost": {
                        "description": "Стоимость",
                        "type": "integer",
                        "format": "integer"
                    },
                    "payment": {
                        "description": "ID тип оплаты",
                        "type": "integer",
                        "format": "integer"
                    },
                    "worksheet": {
                        "description": "ID РЛ",
                        "type": "integer",
                        "format": "integer"
                    },
                    "simple": {
                        "description": "Упрощнный растет",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "close": {
                        "description": "Расторжение",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "number": {
                        "description": "Номр договора",
                        "type": "string",
                        "format": "string"
                    },
                    "begin_at": {
                        "description": "Дата начала",
                        "type": "string",
                        "format": "string"
                    },
                    "register_at": {
                        "description": "Дата оформления",
                        "type": "string",
                        "format": "string"
                    },
                    "decorator": {
                        "description": "ID оформителя",
                        "type": "integer",
                        "format": "integer"
                    },
                    "manager": {
                        "description": "ID продовца",
                        "type": "integer",
                        "format": "integer"
                    },
                    "award": {
                        "description": "Премия",
                        "type": "integer",
                        "format": "integer"
                    },
                    "award_complete": {
                        "description": "Зачисление премии",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "deduction": {
                        "description": "Удержание",
                        "type": "boolean",
                        "format": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}